executor: remove structs package
