+++ drivers/java/handle.go
@@ -8,12 +8,12 @@ import (
 
 	"github.com/hashicorp/go-hclog"
 	"github.com/hashicorp/go-plugin"
-	"github.com/hashicorp/nomad/drivers/shared/executor/structs"
+	"github.com/hashicorp/nomad/drivers/shared/executor"
 	"github.com/hashicorp/nomad/plugins/drivers"
 )
 
 type taskHandle struct {
-	exec         structs.Executor
+	exec         executor.Executor
 	pid          int
 	pluginClient *plugin.Client
 	logger       hclog.Logger
