+++ drivers/shared/executor/executor_linux.go
@@ -19,7 +19,6 @@ import (
 	multierror "github.com/hashicorp/go-multierror"
 	"github.com/hashicorp/nomad/client/stats"
 	cstructs "github.com/hashicorp/nomad/client/structs"
-	"github.com/hashicorp/nomad/drivers/shared/executor/structs"
 	"github.com/hashicorp/nomad/helper/discover"
 	shelpers "github.com/hashicorp/nomad/helper/stats"
 	"github.com/hashicorp/nomad/helper/uuid"
@@ -62,7 +61,7 @@ func init() {
 // LibcontainerExecutor implements an Executor with the runc/libcontainer api
 type LibcontainerExecutor struct {
 	id      string
-	command *structs.ExecCommand
+	command *ExecCommand
 
 	logger hclog.Logger
 
@@ -74,10 +73,10 @@ type LibcontainerExecutor struct {
 	container      libcontainer.Container
 	userProc       *libcontainer.Process
 	userProcExited chan interface{}
-	exitState      *structs.ProcessState
+	exitState      *ProcessState
 }
 
-func NewExecutorWithIsolation(logger hclog.Logger) structs.Executor {
+func NewExecutorWithIsolation(logger hclog.Logger) Executor {
 	logger = logger.Named("isolated_executor")
 	if err := shelpers.Init(); err != nil {
 		logger.Error("unable to initialize stats", "error", err)
@@ -93,7 +92,7 @@ func NewExecutorWithIsolation(logger hclog.Logger) structs.Executor {
 }
 
 // Launch creates a new container in libcontainer and starts a new process with it
-func (l *LibcontainerExecutor) Launch(command *structs.ExecCommand) (*structs.ProcessState, error) {
+func (l *LibcontainerExecutor) Launch(command *ExecCommand) (*ProcessState, error) {
 	l.logger.Info("launching command", "command", command.Cmd, "args", strings.Join(command.Args, " "))
 	// Find the nomad executable to launch the executor process with
 	bin, err := discover.NomadExecutable()
@@ -102,7 +101,7 @@ func (l *LibcontainerExecutor) Launch(command *structs.ExecCommand) (*structs.Pr
 	}
 
 	if command.Resources == nil {
-		command.Resources = &structs.Resources{}
+		command.Resources = &Resources{}
 	}
 
 	l.command = command
@@ -207,7 +206,7 @@ func (l *LibcontainerExecutor) Launch(command *structs.ExecCommand) (*structs.Pr
 	go l.pidCollector.collectPids(l.userProcExited, l.getAllPids)
 	go l.wait()
 
-	return &structs.ProcessState{
+	return &ProcessState{
 		Pid:      pid,
 		ExitCode: -1,
 		Time:     time.Now(),
@@ -232,7 +231,7 @@ func (l *LibcontainerExecutor) getAllPids() (map[int]*nomadPid, error) {
 }
 
 // Wait waits until a process has exited and returns it's exitcode and errors
-func (l *LibcontainerExecutor) Wait(ctx context.Context) (*structs.ProcessState, error) {
+func (l *LibcontainerExecutor) Wait(ctx context.Context) (*ProcessState, error) {
 	select {
 	case <-ctx.Done():
 		return nil, ctx.Err()
@@ -252,7 +251,7 @@ func (l *LibcontainerExecutor) wait() {
 			ps = exitErr.ProcessState
 		} else {
 			l.logger.Error("failed to call wait on user process", "error", err)
-			l.exitState = &structs.ProcessState{Pid: 0, ExitCode: 0, Time: time.Now()}
+			l.exitState = &ProcessState{Pid: 0, ExitCode: 0, Time: time.Now()}
 			return
 		}
 	}
@@ -270,7 +269,7 @@ func (l *LibcontainerExecutor) wait() {
 		}
 	}
 
-	l.exitState = &structs.ProcessState{
+	l.exitState = &ProcessState{
 		Pid:      ps.Pid(),
 		ExitCode: exitCode,
 		Signal:   signal,
@@ -332,13 +331,13 @@ func (l *LibcontainerExecutor) Shutdown(signal string, grace time.Duration) erro
 }
 
 // UpdateResources updates the resource isolation with new values to be enforced
-func (l *LibcontainerExecutor) UpdateResources(resources *structs.Resources) error {
+func (l *LibcontainerExecutor) UpdateResources(resources *Resources) error {
 	return nil
 }
 
 // Version returns the api version of the executor
-func (l *LibcontainerExecutor) Version() (*structs.ExecutorVersion, error) {
-	return &structs.ExecutorVersion{Version: ExecutorVersionLatest}, nil
+func (l *LibcontainerExecutor) Version() (*ExecutorVersion, error) {
+	return &ExecutorVersion{Version: ExecutorVersionLatest}, nil
 }
 
 // Stats returns the resource statistics for processes managed by the executor
@@ -458,7 +457,7 @@ func (l *LibcontainerExecutor) handleExecWait(ch chan *waitResult, process *libc
 	ch <- &waitResult{ps, err}
 }
 
-func configureCapabilities(cfg *lconfigs.Config, command *structs.ExecCommand) {
+func configureCapabilities(cfg *lconfigs.Config, command *ExecCommand) {
 	// TODO: allow better control of these
 	cfg.Capabilities = &lconfigs.Capabilities{
 		Bounding:    allCaps,
@@ -470,7 +469,7 @@ func configureCapabilities(cfg *lconfigs.Config, command *structs.ExecCommand) {
 
 }
 
-func configureIsolation(cfg *lconfigs.Config, command *structs.ExecCommand) {
+func configureIsolation(cfg *lconfigs.Config, command *ExecCommand) {
 	defaultMountFlags := syscall.MS_NOEXEC | syscall.MS_NOSUID | syscall.MS_NODEV
 
 	// set the new root directory for the container
@@ -536,7 +535,7 @@ func configureIsolation(cfg *lconfigs.Config, command *structs.ExecCommand) {
 	}
 }
 
-func configureCgroups(cfg *lconfigs.Config, command *structs.ExecCommand) error {
+func configureCgroups(cfg *lconfigs.Config, command *ExecCommand) error {
 
 	// If resources are not limited then manually create cgroups needed
 	if !command.ResourceLimits {
@@ -602,7 +601,7 @@ func configureBasicCgroups(cfg *lconfigs.Config) error {
 	return nil
 }
 
-func newLibcontainerConfig(command *structs.ExecCommand) *lconfigs.Config {
+func newLibcontainerConfig(command *ExecCommand) *lconfigs.Config {
 	cfg := &lconfigs.Config{
 		Cgroups: &lconfigs.Cgroup{
 			Resources: &lconfigs.Resources{
